requirements.txt
anyio
aspose-words
beautifulsoup4
cachetools
certifi
pymongo
charset-normalizer
click
dnspython
email-validator
fastapi
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
h11
httpcore
httplib2
httptools
httpx
pytest
idna
itsdangerous
Jinja2
MarkupSafe
orjson
protobuf
pyasn1
pyasn1-modules
pydantic
pyparsing
PyPDF2
python-dotenv
python-multipart
PyYAML
requests
rsa
six
sniffio
soupsieve
starlette
tqdm
typing_extensions
ujson
uritemplate
urllib3
uvicorn
# uvloop
watchfiles
websockets
beautifulsoup4
bs4
